# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: HRMS API
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /health:
    get:
      summary: Service health
      responses:
        '200':
          description: OK

  /api/v1/auth/register:
    post:
      summary: Register tenant + admin
      requestBody:
        required: true
      responses:
        '201':
          description: Created

  /api/v1/auth/login:
    post:
      summary: Login
      requestBody:
        required: true
      responses:
        '200':
          description: Access and refresh tokens

  /api/v1/auth/refresh:
    post:
      summary: Refresh access token
      requestBody:
        required: true
      responses:
        '200':
          description: Refreshed

  /api/v1/auth/me:
    get:
      summary: Current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User payload

  /api/v1/auth/google/start:
    get:
      summary: Start Google OAuth
      responses:
        '200':
          description: OAuth URL

  /api/v1/auth/google/callback:
    get:
      summary: Google OAuth callback
      responses:
        '200':
          description: Callback processed

  /api/v1/employees:
    get:
      summary: List employees
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Employee list
    post:
      summary: Create employee
      security:
        - bearerAuth: []
      requestBody:
        required: true
      responses:
        '201':
          description: Employee created

  /api/v1/employees/{id}:
    patch:
      summary: Update employee
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Employee updated

  /api/v1/attendance:
    get:
      summary: List attendance logs
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Attendance logs

  /api/v1/attendance/check-in:
    post:
      summary: Attendance check-in
      security:
        - bearerAuth: []
      requestBody:
        required: true
      responses:
        '201':
          description: Check-in created

  /api/v1/attendance/check-out:
    post:
      summary: Attendance check-out
      security:
        - bearerAuth: []
      requestBody:
        required: true
      responses:
        '200':
          description: Check-out updated

  /api/v1/audit-logs:
    get:
      summary: List audit logs
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Audit records

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT